requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
facebook-sdk==3.1.0
